requirements.txt
opencv-python
ncnn
ultralytics
pyserial